requirements.txt
filetype==1.2.0
packaging==23.2
Pillow==10.1.0